info/azure-function/requirements.txt
azure-functions>=1.17.0
azure-storage-blob==12.17.0